+++ src/Boot/src/BootloadManager/BootloadManager.php
@@ -38,17 +38,17 @@ final class BootloadManager implements Container\SingletonInterface
      * ]
      *
      * @param array<class-string>|array<class-string,array<string,mixed>> $classes
-     * @param array<Closure> $staringCallbacks
-     * @param array<Closure> $startedCallbacks
+     * @param array<Closure> $bootingCallbacks
+     * @param array<Closure> $bootedCallbacks
      *
      * @throws \Throwable
      */
-    public function bootload(array $classes, array $staringCallbacks = [], array $startedCallbacks = []): void
+    public function bootload(array $classes, array $bootingCallbacks = [], array $bootedCallbacks = []): void
     {
         $this->container->runScope(
             [self::class => $this],
-            function () use ($classes, $staringCallbacks, $startedCallbacks): void {
-                $this->boot($classes, $staringCallbacks, $startedCallbacks);
+            function () use ($classes, $bootingCallbacks, $bootedCallbacks): void {
+                $this->boot($classes, $bootingCallbacks, $bootedCallbacks);
             }
         );
     }